Score and timer in Game
Score works for decreasing it in the case of an incorrect order.
For a correct sandwich, pass nothing into SandwichTimer.FinishSandwich()
For an incorrect sandwich, send a 0 through to the same function.

The function must be called to increase score when a sandwich is completed

Score.ChangeScore will manually change a score with an integer passed through to it